+++ src/RoundView/RoundAlbumListItem/RoundAlbumListItem.js
@@ -9,8 +9,10 @@ import { Link } from 'react-router-dom';
 
 // Import components
 import HeartbeatChart from './HeartbeatChart';
+import { MemberIconMedium } from '../../MemberIcons/MemberIcons';
 import MissingVotesInfo from './MissingVotesInfo'
 import PickedTrackTable from '../PickedTrackTable/PickedTrackTable';
+
 
 function RoundAlbumListItem({ album, participants, votesPerParticipant }) {
   const showPickedTrackTable = album.tracks !== null && album.tracks.length > 0;
@@ -28,36 +30,45 @@ function RoundAlbumListItem({ album, participants, votesPerParticipant }) {
 
   return (
     <div className={`${styles.RoundAlbumListItem} row`}>
-      <div className="col-sm-4 d-flex flex-column align-items-center">
-        <Link to={'/album/' + album.id} style={{textDecoration: 'none'}}>
-          <div className={`${styles.postedAlbumIcon} mb-3`} style={{backgroundImage: 'url(' + album.imageUrl + ')'}}>
-            <div className={styles.posterIcon} style={{backgroundColor: poster.color}}>
-              <p className="m-0">{poster.firstName[0] + poster.lastName[0]}</p>
-            </div>
-            <div className={styles.iconOverlay}></div>
+      <div className={`${styles.albumInfoArea} col-lg-4`}>
+        <div className="d-flex flex-column align-items-center">
+          <div className={`${styles.posterHeader} d-flex align-items-center mb-3`}>
+            <MemberIconMedium key={poster.id} member={poster} />
+            <p className="m-0">posted</p>
           </div>
-        </Link>
 
-        <Link to={'/album/' + album.id} style={{textDecoration: 'none'}}>
-          <h2 className="text-center">{album.title}</h2>
-        </Link>
+          <Link to={'/album/' + album.id} style={{textDecoration: 'none'}}>
+            <img className={`${styles.postedAlbumImg}`} src={album.imageUrl} alt={`${album.title} cover art`} />
+          </Link>
+        </div>
 
-        <h3 className="text-center mb-3">{album.artists.join(', ')}</h3>
+        <div className="flex-grow-1">
+          <div className={styles.albumDescription}>
+            <Link to={'/album/' + album.id} style={{textDecoration: 'none'}}>
+              <h2 className="text-center">{album.title}</h2>
+            </Link>
 
-        {
-          isComplete
-          ? <p className="mb-2" title="Vote overlap score. There is 100% overlap if everyone votes for the same songs, and 0% overlap if everyone votes for different songs."><FontAwesomeIcon icon={faHandshake} /> {alignmentPercentage}%</p>
-          : <MissingVotesInfo unfinishedParticipants={unfinishedParticipants}/>
-        }
+            <h3 className="text-center mb-2">{album.artists.join(', ')}</h3>
 
-        {
-          isComplete
-          ? <div className="mb-4"><HeartbeatChart album={album} /></div>
-          : null
-        }
+            <div className={`${styles.albumStatsArea} mb-0`}>
+              {
+                isComplete
+                ? <div className="m-0"><HeartbeatChart album={album} title="Hello" /></div>
+                : null
+              }
+
+              {
+                isComplete
+                ? <p className="m-0" title="Vote overlap score. There is 100% overlap if everyone votes for the same songs, and 0% overlap if everyone votes for different songs."><FontAwesomeIcon icon={faHandshake} /> {alignmentPercentage}%</p>
+                : <MissingVotesInfo unfinishedParticipants={unfinishedParticipants}/>
+              }
+            </div>
+          </div>
+        </div>
+
       </div>
 
-      <div className="col-sm-8">
+      <div className="col-lg-8">
         {
           showPickedTrackTable
           ? <PickedTrackTable
@@ -96,3 +107,7 @@ function createParticipantVotes(participants, tracks) {
 }
 
 export default RoundAlbumListItem;
+
+/*
+
+*/