+++ src/RoundView/RoundHeader/RoundHeader.js
@@ -15,16 +15,19 @@ function RoundHeader({round, participants}) {
   const dayNumberStr = (dayNumber === 1) ? '1 day' : dayNumber + ' days';
 
   return (
-    <div className={`${styles.RoundHeader} mt-3 d-flex justify-content-between`}>
-      <div>
-        <h1 className="m-0">Round {round.number}</h1>
-        <h2 className="m-0">{startDateStr} - {endDateStr} ({dayNumberStr})</h2>
-        {
-          round.description ? <p className={`${styles.roundDescription} mt-4`}>{ round.description}</p> : null
-        }
+    <div className={`${styles.RoundHeader} mt-3`}>
+      <div className="d-flex justify-content-between">
+        <div>
+          <h1 className="m-0">Round {round.number}</h1>
+          <h2 className="m-0">{startDateStr} - {endDateStr} ({dayNumberStr})</h2>
+        </div>
+
+        <RoundParticipantsKey participants={participants} />
       </div>
 
-      <RoundParticipantsKey participants={participants} />
+      {
+        round.description ? <p className={`${styles.roundDescription} mt-4`}>{ round.description}</p> : null
+      }
     </div>
   );
 }