+++ src/RoundView/RoundParticipantsKey/RoundParticipantsKey.js
@@ -3,19 +3,22 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import styles from './RoundParticipantsKey.module.css';
 
 // Import components
-import { MemberIconSmall } from '../../MemberIcons/MemberIcons';
+import { MemberIconMedium } from '../../MemberIcons/MemberIcons';
 
 function RoundParticipantsKey({participants}) {
   return (
-    <div className={`${styles.RoundParticipantsKey} d-flex flex-column align-items-end`}>
-    {
-      participants.map((participant) => (
-        <div className={`${styles.participantListItem} mb-2`} key={participant.id}>
-          <p className={styles.participantName}>{participant.firstName} {participant.lastName}</p>
-          <MemberIconSmall member={participant} />
-        </div>
-      ))
-    }
+    <div className={styles.RoundParticipantsKey}>
+      <h2>Participants</h2>
+
+      <div className={styles.participantsList}>
+        {
+          participants.map((participant) => (
+            <div className={`${styles.participantListItem}`} key={participant.id}>
+              <MemberIconMedium member={participant} />
+            </div>
+          ))
+        }
+      </div>
     </div>
   );
 }